v1.5

+++ MTGAHelper.Entity/OutputLogParsing/MythicRatingUpdatedRaw.cs
@@ -4,8 +4,8 @@
     {
         public float oldMythicPercentile { get; set; }
         public float newMythicPercentile { get; set; }
-        public float oldMythicLeaderboardPlacement { get; set; }
-        public float newMythicLeaderboardPlacement { get; set; }
+        public int oldMythicLeaderboardPlacement { get; set; }
+        public int newMythicLeaderboardPlacement { get; set; }
         public string context { get; set; }
     }
 }
